third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/import-scripts-updated-flag.https-expected.txt

This is a testharness.js-based test.
PASS initialize global state
PASS import script previously imported at worker evaluation time
PASS import script previously imported at worker install time
FAIL import script not previously imported assert_equals: expected (string) "TypeError" but got (object) null
PASS Tests for importScripts: import scripts updated flag
Harness: the test ran to completion.
